# api/openapi.yaml
openapi: 3.0.3
info:
  title: CertForge API
  version: 0.1.0
  description: |
    CertForge REST API. All paths are relative to /api/v1.

    Authentication: the `cf_session` cookie set by POST /auth/login or POST /setup/complete.
    POST, PUT, PATCH, and DELETE with a session must send the `X-CSRF-Token` header with
    `csrfToken` from GET /auth/me. Request bodies must be `application/json`.

    Errors are RFC 9457 application/problem+json documents (see the Problem schema and
    components.responses): 400, 401, 403 (missing permission or CSRF failure), 404, 409,
    413, 415, 422, 429 (login rate limit, with Retry-After), 500, 502 (upstream CA or
    identity provider), 503 (busy, with Retry-After).
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /setup/status:
    get:
      operationId: getSetupStatus
      tags: [setup]
      summary: First-run status
      description: Whether the first-run setup wizard still needs to run. Public.
      security: []
      responses:
        "200":
          description: Setup status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupStatus"
        "500": {$ref: "#/components/responses/InternalError"}
  /setup/complete:
    post:
      operationId: completeSetup
      tags: [setup]
      summary: Complete first-run setup
      description: Sets the local admin password, stores the base URL, creates the first org, and logs the admin in. Works once; later calls get 409. Public.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "200":
          description: Setup done; session cookie set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400": {$ref: "#/components/responses/BadRequest"}
        "409": {$ref: "#/components/responses/Conflict"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
        "500": {$ref: "#/components/responses/InternalError"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in as the local admin
      description: Break-glass local admin login. Sets the cf_session cookie and revokes the user's other sessions. Rate limited per client address (429 with Retry-After).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in; session cookie set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400": {$ref: "#/components/responses/BadRequest"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
        "429": {$ref: "#/components/responses/TooManyRequests"}
        "500": {$ref: "#/components/responses/InternalError"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
...
  /auth/methods:
    get:
      operationId: getAuthMethods
      tags: [auth]
      summary: Available sign-in methods
      description: Which sign-in methods the login page offers. Public.
      security: []
      responses:
        "200":
          description: Sign-in methods.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthMethods"}
        "500": {$ref: "#/components/responses/InternalError"}
  /auth/oidc/start:
    get:
      operationId: startOidcLogin
      tags: [auth]
      summary: Start single sign-on
      description: Redirects the browser to the identity provider (auth code with PKCE) and sets a short-lived signed state cookie. Redirects to /login?error=oidc_disabled when single sign-on is off. Public.
      security: []
      parameters:
        - name: next
          in: query
          required: false
          description: Same-origin path to open after sign-in; anything else becomes /.
          schema: {type: string}
      responses:
        "302":
          description: Redirect to the identity provider or back to /login.
          headers:
            Location:
              description: Redirect target.
              schema: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
        "500": {$ref: "#/components/responses/InternalError"}
  /auth/oidc/callback:
    get:
      operationId: oidcCallback
      tags: [auth]
      summary: Single sign-on callback
      description: The identity provider returns here. On success creates or updates the user by (issuer, subject), records groups, starts a session (revoking the user's others), and redirects to next. On failure redirects to /login?error= one of oidc_state, oidc_denied, oidc_failed, oidc_disabled, user_disabled. Public; rate limited per client address.
      security: []
      parameters:
        - {name: code, in: query, required: false, description: Authorization code., schema: {type: string}}
        - {name: state, in: query, required: false, description: State echoed by the provider., schema: {type: string}}
        - {name: error, in: query, required: false, description: Error code from the provider., schema: {type: string}}
        - {name: error_description, in: query, required: false, description: Provider error text; logged only., schema: {type: string}}
      responses:
        "302":
          description: Redirect to next or to /login?error=.
          headers:
            Location:
              description: Redirect target.
              schema: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
        "429": {$ref: "#/components/responses/TooManyRequests"}
        "500": {$ref: "#/components/responses/InternalError"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cf_session
      description: Session cookie from POST /auth/login.
...
  responses:
    BadRequest:
      description: Malformed request, query parameter, or JSON body.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    Conflict:
      description: Conflicts with current state, for example a duplicate name, existing dependents, or nothing to confirm.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    PayloadTooLarge:
      description: Request body exceeds 1 MiB.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    UnsupportedMediaType:
      description: Request body is not application/json.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    UnprocessableEntity:
      description: The body parsed but failed validation; the title or detail names the field.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    TooManyRequests:
      description: Too many login attempts from this address.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: {type: integer}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    InternalError:
      description: Unexpected server error; details are in the server log.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    ServiceUnavailable:
      description: Temporarily at capacity.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: {type: integer}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details, sent with content type application/problem+json.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: Problem type URI; about:blank for generic HTTP errors.
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Explanation specific to this occurrence.
    Org:
      type: object
      description: An organization; most resources are scoped to one.
      required: [id, slug, name]
      properties:
        id:
          type: string
          format: uuid
          description: Org id.
        slug:
          type: string
          description: URL-safe identifier, lowercase letters, digits, and hyphens.
        name:
          type: string
          description: Display name.
...
    LoginRequest:
      type: object
      description: Local admin credentials.
      required: [password]
      properties:
        password:
          type: string
          format: password
          description: Local admin password.
    AuthMethods:
      type: object
      description: Sign-in methods offered on the login page.
      required: [oidcEnabled, localEnabled]
      properties:
        oidcEnabled: {type: boolean, description: Single sign-on is enabled and configured.}
        localEnabled: {type: boolean, description: A local (break-glass) admin exists.}
...
    User:
      type: object
      description: A CertForge user.
      required: [id, displayName, localAdmin]
      properties:
        id:
          type: string
          format: uuid
          description: User id.
        displayName:
          type: string
          description: Name shown in the UI.
        localAdmin:
          type: boolean
          description: True for the break-glass local admin.
    Role:
      type: string
      enum: [admin, org-admin, operator, viewer, auditor]
      description: A role. admin is everything; org-admin everything in its org except global-only actions; operator certificates, credentials, accounts, clients, issue and renew; viewer read-only; auditor viewer plus the audit log.
    MeBinding:
      type: object
      description: One role held by the signed-in principal.
      required: [role, orgId]
      properties:
        role:
          $ref: "#/components/schemas/Role"
        orgId:
          type: string
          format: uuid
          nullable: true
          description: The org the role applies to; null for a global binding.
    Me:
      type: object
      description: The signed-in principal.
      required: [user, roles, bindings, orgs, csrfToken]
      properties:
        user:
          description: The signed-in user.
          allOf:
            - $ref: "#/components/schemas/User"
        roles:
          type: array
          description: Distinct role names bound to the user.
          items:
            type: string
        bindings:
          type: array
          description: The roles behind `roles`, with their scope. The UI mirrors authz.Can from these.
          items: {$ref: "#/components/schemas/MeBinding"}
        orgs:
          type: array
          description: Orgs the user can see.
          items:
            $ref: "#/components/schemas/Org"
        csrfToken:
          type: string
          description: Send in the X-CSRF-Token header on every POST, PUT, PATCH, and DELETE.
...
    SetupStatus:
      type: object
      description: First-run state.
      required: [needsSetup]
      properties:
        needsSetup:
          type: boolean
          description: True until POST /setup/complete succeeds.
    SetupRequest:
      type: object
      description: First-run wizard input.
      required: [adminPassword, orgName, orgSlug, baseUrl]
      properties:
        adminPassword:
          type: string
          format: password
          minLength: 12
          description: New local admin password, at least 12 characters.
        orgName:
          type: string
          minLength: 1
          maxLength: 100
          description: Display name of the first org.
        orgSlug:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{0,62}$"
          description: URL-safe identifier of the first org.
        baseUrl:
          type: string
          description: Public URL of this CertForge instance, for example https://certs.example.com.
...